Improve connection opening speed & blacklist servers that are DOWN
Loop through servers, and if connection is in best possible state, return and use it immediately,
instead of looping through other servers, before finding best connection.
This should always be faster way to resolve best connection, than solution there was previously.

Also blacklist servers that are down, so that they are not retried all the time